fix/バックエンドのディレクトリ構成修正
