+++ src/core/profile/repositories/infra.ts
@@ -1,116 +1,16 @@
 import { pool } from "@/lib/db";
 import {
-  ProfileData,
   AnsweredData,
   Subject,
   Answer,
   CardAllDataRaw,
   TestSubmissionData,
   TestSubmissionResult,
-  FollowUser,
   TestResult,
-  ProfileUpdateResponse,
 } from "../type";
+import { FollowUser } from "@/type/followType";
 
 export class ProfileRepository {
-  // プロフィールデータの取得
-  async fetchProfileData(): Promise<ProfileData> {
-    const query = `
-      SELECT 
-        u.id, 
-        u.user_name, 
-        u.memo,
-        u.targetUniversity_1,
-        u.targetUniversity_2,
-        u.targetUniversity_3,
-        u.thumbnail_url as "thumbnailUrl"
-      FROM 
-        users u
-      WHERE 
-        u.id = 1
-    `;
-
-    try {
-      const result = await pool.query(query);
-
-      if (result.rows.length === 0) {
-        throw new Error("プロフィールデータが見つかりません");
-      }
-
-      // 志望大学データを配列に変換
-      const targetUniversities = [];
-      if (result.rows[0].targetuniversity_1)
-        targetUniversities.push(result.rows[0].targetuniversity_1);
-      if (result.rows[0].targetuniversity_2)
-        targetUniversities.push(result.rows[0].targetuniversity_2);
-      if (result.rows[0].targetuniversity_3)
-        targetUniversities.push(result.rows[0].targetuniversity_3);
-
-      const profileData: ProfileData = {
-        userName: result.rows[0].user_name,
-        targetUniversities: targetUniversities,
-        memo: result.rows[0].memo,
-        thumbnailUrl: result.rows[0].thumbnailUrl,
-      };
-
-      return profileData;
-    } catch (error) {
-      console.error("プロフィールデータ取得エラー:", error);
-      throw error;
-    }
-  }
-
-  // プロフィール情報の更新
-  async updateProfileData(data: {
-    userId: number;
-    userName: string;
-    targetUniversities: string[];
-    memo: string;
-    thumbnailUrl?: string;
-  }): Promise<ProfileUpdateResponse> {
-    try {
-      // トランザクション開始
-      await pool.query("BEGIN");
-
-      // usersテーブルを更新
-      await pool.query(
-        `UPDATE users
-       SET 
-        user_name = $1,
-        targetUniversity_1 = $2,
-        targetUniversity_2 = $3,
-        targetUniversity_3 = $4,
-        memo = $5,
-        thumbnail_url = $6,
-        updated_at = CURRENT_TIMESTAMP
-       WHERE 
-        id = $7`,
-        [
-          data.userName,
-          data.targetUniversities[0] || null,
-          data.targetUniversities[1] || null,
-          data.targetUniversities[2] || null,
-          data.memo,
-          data.thumbnailUrl || null,
-          data.userId,
-        ]
-      );
-
-      // トランザクションコミット
-      await pool.query("COMMIT");
-
-      return { success: true };
-    } catch (error) {
-      // エラー発生時はロールバック
-      await pool.query("ROLLBACK");
-      console.error("プロフィール更新エラー:", error);
-      return {
-        success: false,
-        error: error instanceof Error ? error.message : String(error),
-      };
-    }
-  }
-
   // testsテーブルからデータを取得するメソッド
   async fetchCardAllDatasRaw(): Promise<CardAllDataRaw[]> {
     const query = `
@@ -723,123 +623,6 @@ export class ProfileRepository {
         success: false,
         error: error instanceof Error ? error.message : String(error),
       };
-    }
-  }
-
-  // テスト目標の設定
-  async saveTestTargets(data: {
-    userId: number;
-    subject: Subject;
-    targetPercentage: number;
-    targetMonth: string;
-    targetMemo?: string;
-    targetSectionPercentages: Record<number, number>;
-  }): Promise<{ success: boolean; error?: string }> {
-    try {
-      // トランザクション開始
-      await pool.query("BEGIN");
-
-      await pool.query(
-        `INSERT INTO tests_target
-          (user_id, subject, target_percentage, target_month, target_memo,
-          target_percentage_section1, target_percentage_section2, target_percentage_section3,
-          target_percentage_section4, target_percentage_section5, target_percentage_section6)
-         VALUES
-          ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11)
-         ON CONFLICT (user_id, subject, target_month)
-         DO UPDATE SET
-          target_percentage = $3,
-          target_memo = $5,
-          target_percentage_section1 = $6,
-          target_percentage_section2 = $7,
-          target_percentage_section3 = $8,
-          target_percentage_section4 = $9,
-          target_percentage_section5 = $10,
-          target_percentage_section6 = $11,
-          updated_at = CURRENT_TIMESTAMP`,
-        [
-          data.userId,
-          data.subject,
-          data.targetPercentage,
-          data.targetMonth,
-          data.targetMemo || "",
-          data.targetSectionPercentages[1] || null,
-          data.targetSectionPercentages[2] || null,
-          data.targetSectionPercentages[3] || null,
-          data.targetSectionPercentages[4] || null,
-          data.targetSectionPercentages[5] || null,
-          data.targetSectionPercentages[6] || null,
-        ]
-      );
-
-      // トランザクションコミット
-      await pool.query("COMMIT");
-
-      return { success: true };
-    } catch (error) {
-      // エラー発生時はロールバック
-      await pool.query("ROLLBACK");
-      console.error("テスト目標設定エラー:", error);
-      return {
-        success: false,
-        error: error instanceof Error ? error.message : String(error),
-      };
-    }
-  }
-
-  // テスト目標の取得
-  async fetchTestTarget(
-    subject: string,
-    targetMonth: string
-  ): Promise<{
-    targetPercentage: number;
-    targetMonth: string;
-    targetMemo?: string;
-    targetSectionPercentages: Record<number, number>;
-  } | null> {
-    const query = `
-      SELECT 
-        target_percentage,
-        target_month,
-        target_memo,
-        json_build_object(
-          1, target_percentage_section1,
-          2, target_percentage_section2,
-          3, target_percentage_section3,
-          4, target_percentage_section4,
-          5, target_percentage_section5,
-          6, target_percentage_section6
-        ) AS target_section_percentages
-      FROM 
-        tests_target
-      WHERE 
-        subject = $1 AND user_id = 1 AND target_month = $2
-      LIMIT 1
-    `;
-
-    try {
-      const result = await pool.query(query, [subject, targetMonth]);
-
-      if (result.rows.length === 0) {
-        return null;
-      }
-
-      const row = result.rows[0];
-
-      // NULL値を除去するためにオブジェクトをフィルタリング
-      const targetSectionPercentages = this.filterNullValues(
-        row.target_section_percentages
-      );
-
-      return {
-        targetPercentage: row.target_percentage,
-        targetMonth: row.target_month,
-        targetMemo: row.target_memo,
-        targetSectionPercentages,
-      };
-    } catch (error) {
-      console.error("テスト目標取得エラー:", error);
-      throw error;
     }
   }
 }
