+++ src/type/newsType.ts
