+++ src/type/userType.ts
@@ -1,0 +1,18 @@
+export interface UserData {
+  userName: string;
+  targetUniversities: string[];
+  memo: string;
+  thumbnailUrl: string;
+}
+
+export interface UserUpdateData {
+  userName: string;
+  targetUniversities: string[];
+  memo: string;
+  thumbnailUrl?: string;
+}
+
+export interface UserUpdateResponse {
+  success: boolean;
+  error?: string;
+}
